+++ app/src/main/res/layout/account_by_admin_activity.xml
@@ -42,7 +42,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:textAppearance="@style/TextAppearance.AppCompat.Small"
-                android:text="@string/email"/>
+               />
         </LinearLayout>
         <LinearLayout
             android:layout_width="0dp"
